python-requirements.txt
https://www.djangoproject.com/download/1.7c1/tarball/
selenium
psycopg2
django-nose
watchdog
ipython
sh

